+++ src/main/java/com/codetaylor/mc/pyrotech/modules/plugin/patchouli/ModulePluginPatchouli.java
@@ -1,12 +1,8 @@
 package com.codetaylor.mc.pyrotech.modules.plugin.patchouli;
 
 import com.codetaylor.mc.athenaeum.module.ModuleBase;
-import com.codetaylor.mc.athenaeum.registry.Registry;
 import com.codetaylor.mc.pyrotech.ModPyrotech;
-import com.codetaylor.mc.pyrotech.modules.plugin.patchouli.event.PlayerJoinEventHandler;
-import com.codetaylor.mc.pyrotech.modules.plugin.patchouli.init.EntityInitializer;
-import com.codetaylor.mc.pyrotech.modules.plugin.patchouli.init.ItemInitializer;
-import com.codetaylor.mc.pyrotech.modules.plugin.patchouli.item.ItemBook;
+import com.codetaylor.mc.pyrotech.modules.plugin.event.PlayerJoinEventHandler;
 import com.codetaylor.mc.pyrotech.packer.PackAPI;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.IResource;
@@ -15,7 +11,6 @@ import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.common.registry.GameRegistry;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import vazkii.patchouli.api.PatchouliAPI;
@@ -36,9 +31,6 @@ public class ModulePluginPatchouli
   public ModulePluginPatchouli() {
 
     super(0, MOD_ID);
-
-    this.setRegistry(new Registry(MOD_ID, CREATIVE_TAB));
-    this.enableAutoRegistry();
 
     MinecraftForge.EVENT_BUS.register(new PlayerJoinEventHandler());
   }
@@ -119,29 +111,5 @@ public class ModulePluginPatchouli
           return null;
         }
     );
-  }
-
-  @Override
-  public void onRegister(Registry registry) {
-
-    ItemInitializer.onRegister(registry);
-    EntityInitializer.onRegister(registry);
-  }
-
-  @Override
-  public void onClientRegister(Registry registry) {
-
-    ItemInitializer.onClientRegister(registry);
-  }
-
-  @GameRegistry.ObjectHolder(ModulePluginPatchouli.MOD_ID)
-  public static class Items {
-
-    @GameRegistry.ObjectHolder(ItemBook.NAME)
-    public static final ItemBook BOOK;
-
-    static {
-      BOOK = null;
-    }
   }
 }