+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/init/recipe/CampfireRecipesAdd.java
@@ -1,5 +1,7 @@
 package com.codetaylor.mc.pyrotech.modules.tech.basic.init.recipe;
 
+import com.codetaylor.mc.pyrotech.ModPyrotech;
+import com.codetaylor.mc.pyrotech.modules.core.ModuleCore;
 import com.codetaylor.mc.pyrotech.modules.plugin.patchouli.ModulePluginPatchouli;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.ModuleTechBasic;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.recipe.CampfireRecipe;
@@ -12,9 +14,11 @@ public class CampfireRecipesAdd {
 
   public static void apply(IForgeRegistry<CampfireRecipe> registry) {
 
-    registry.register(new CampfireRecipe(
-        new ItemStack(ModulePluginPatchouli.Items.BOOK),
-        Ingredient.fromStacks(new ItemStack(Items.BOOK))
-    ).setRegistryName(ModuleTechBasic.MOD_ID, "book"));
+    if (ModPyrotech.INSTANCE.isModuleEnabled(ModulePluginPatchouli.class)) {
+      registry.register(new CampfireRecipe(
+          new ItemStack(ModuleCore.Items.BOOK),
+          Ingredient.fromStacks(new ItemStack(Items.BOOK))
+      ).setRegistryName(ModuleTechBasic.MOD_ID, "book"));
+    }
   }
 }